release: v0.6.1
- default wandb training mode -> disabled (CLI --mode + Trainer.train); pass online to log
- relocate the simplify=='sympy' helper to a dependency-light leaf (utils/sympy_timeout); model +
  skeleton pool import from there (internal; no behaviour change)
- defensive: simplify=='sympy' raises a clear 'pip install flash-ansr[sympy]' if sympy is ever absent
  (in practice sympy ships transitively with torch)
